+++ mysql_class.py
@@ -42,6 +42,7 @@ __version__ = version.__version__
 # Global
 KEY1 = "pass"
 KEY2 = "wd"
+SHOW = "show "
 
 
 def fetch_global_var(server, var):
@@ -79,7 +80,9 @@ def fetch_sys_var(server, var, **kwargs):
 
     """
 
-    cmd = "show " + kwargs.get("level", "session") + " variables like %s"
+    global SHOW
+
+    cmd = SHOW + kwargs.get("level", "session") + " variables like %s"
 
     return server.vert_sql(cmd, (var,))
 
@@ -125,7 +128,12 @@ def show_slave_hosts(server):
 
     """
 
-    return server.col_sql("show slave hosts")
+    global SHOW
+
+    # Semantic change in MySQL 8.0.22
+    slaves = "replicas" if server.version >= (8, 0, 22) else "slave hosts"
+
+    return server.col_sql(SHOW + slaves)
 
 
 def show_slave_stat(server):
@@ -140,7 +148,12 @@ def show_slave_stat(server):
 
     """
 
-    return server.col_sql("show slave status")
+    global SHOW
+
+    # Semantic change in MySQL 8.0.22
+    slave = "replica" if server.version >= (8, 0, 22) else "slave"
+
+    return server.col_sql(SHOW + slave + " status")
 
 
 def slave_start(server):
@@ -154,7 +167,10 @@ def slave_start(server):
 
     """
 
-    server.cmd_sql("start slave")
+    # Semantic change in MySQL 8.0.22
+    slave = "replica" if server.version >= (8, 0, 22) else "slave"
+
+    server.cmd_sql("start " + slave)
 
 
 def slave_stop(server):
@@ -168,7 +184,10 @@ def slave_stop(server):
 
     """
 
-    server.cmd_sql("stop slave")
+    # Semantic change in MySQL 8.0.22
+    slave = "replica" if server.version >= (8, 0, 22) else "slave"
+
+    server.cmd_sql("stop " + slave)
 
 
 class Position(collections.namedtuple("Position", "file, pos")):
@@ -512,6 +531,7 @@ class Server(object):
         set_pass_config
         setup_ssl
         set_ssl_config
+        set_tls_config
 
     """
 
@@ -525,7 +545,7 @@ class Server(object):
             (input) name -> Name of the MySQL server.
             (input) server_id -> Server's ID.
             (input) sql_user -> SQL user's name.
-            (input) sql_pass -> SQL user's password.
+            (input) sql_pass -> SQL user's pswd.
             (input) os_type -> Machine operating system type class instance.
             (input) kwargs:
                 extra_def_file -> Location of extra defaults file.
@@ -567,6 +587,10 @@ class Server(object):
         self.ssl_verify_id = kwargs.get("ssl_verify_id", False)
         self.ssl_verify_cert = kwargs.get("ssl_verify_cert", False)
         self.set_ssl_config()
+
+        # TLS configuration settings
+        self.tls_versions = kwargs.get("tls_versions", list())
+        self.set_tls_config()
 
         # SQL connection handler.
         self.conn = None
@@ -840,12 +864,16 @@ class Server(object):
 
         """
 
+        # Semantic change in MySQL 8.0.26
+        master = "source" if self.version >= (8, 0, 26) else "master"
+        slave = "replica" if self.version >= (8, 0, 26) else "slave"
+
         self.log_bin = fetch_sys_var(self, "log_bin")["log_bin"]
         self.read_only = fetch_sys_var(self, "read_only")["read_only"]
         self.log_slv_upd = fetch_sys_var(
-            self, "log_slave_updates")["log_slave_updates"]
+            self, "log_" + slave + "_updates")["log_" + slave + "_updates"]
         self.sync_mst = fetch_sys_var(
-            self, "sync_master_info")["sync_master_info"]
+            self, "sync_" + master + "_info")["sync_" + master + "_info"]
         self.sync_relay = fetch_sys_var(
             self, "sync_relay_log")["sync_relay_log"]
         self.sync_rly_info = fetch_sys_var(
@@ -877,11 +905,15 @@ class Server(object):
 
         """
 
+        # Semantic change in MySQL 8.0.26
+        master = "source" if self.version >= (8, 0, 26) else "master"
+        slave = "replica" if self.version >= (8, 0, 26) else "slave"
+
         return {"log_bin": self.log_bin,
                 "sync_relay_log": self.sync_relay,
                 "read_only": self.read_only,
-                "sync_master_info": self.sync_mst,
-                "log_slave_updates": self.log_slv_upd,
+                "sync_" + master + "_info": self.sync_mst,
+                "log_" + slave + "_updates": self.log_slv_upd,
                 "sync_relay_log_info": self.sync_rly_info}
 
     def upd_log_stats(self):
@@ -1183,6 +1215,22 @@ class Server(object):
                 self.config["ssl_key"] = self.ssl_client_key
                 self.config["ssl_cert"] = self.ssl_client_cert
 
+    def set_tls_config(self):
+
+        """Method:  set_tls_config
+
+        Description:  Append TLS attributes to config.
+
+        Arguments:
+
+        """
+
+        if self.tls_versions and isinstance(self.tls_versions, list):
+            self.config["tls_versions"] = self.tls_versions
+
+        elif self.tls_versions:
+            self.config["tls_versions"] = [self.tls_versions]
+
 
 class Rep(Server):
 
@@ -1202,6 +1250,7 @@ class Rep(Server):
         show_slv_state
         fetch_do_db
         fetch_ign_db
+        verify_srv_id
 
     """
 
@@ -1215,7 +1264,7 @@ class Rep(Server):
             (input) name -> Name of the MySQL server.
             (input) server_id -> Server's ID.
             (input) sql_user -> SQL user's name.
-            (input) sql_pass -> SQL user's password.
+            (input) sql_pass -> SQL user's pswd.
             (input) os_type -> Machine operating system type class instance.
             (input) **kwargs:
                 extra_def_file -> Location of extra defaults file.
@@ -1308,7 +1357,7 @@ class Rep(Server):
 
         var = "server_id"
 
-        return fetch_sys_var(self, var)[var]
+        return int(fetch_sys_var(self, var)[var])
 
     def fetch_do_db(self):
 
@@ -1346,6 +1395,20 @@ class Rep(Server):
 
         return ign_dic
 
+    def verify_srv_id(self):
+
+        """Method:  verify_srv_id
+
+        Description:  Checks to see if the instance configuration file server
+            id matches with the database's server id.
+
+        Arguments:
+            (output) True|False -> If config file id and database id match
+
+        """
+
+        return True if self.server_id == self.get_serv_id() else False
+
 
 class MasterRep(Rep):
 
@@ -1375,7 +1438,7 @@ class MasterRep(Rep):
             (input) name -> Name of the MySQL server.
             (input) server_id -> Server's ID.
             (input) sql_user -> SQL user's name.
-            (input) sql_pass -> SQL user's password.
+            (input) sql_pass -> SQL user's pswd.
             (input) os_type -> Machine operating system type class instance.
             (input) host -> 'localhost' or host name or IP.
             (input) port -> '3306' or port for MySQL.
@@ -1383,7 +1446,7 @@ class MasterRep(Rep):
             (input) **kwargs:
                 extra_def_file -> Location of extra defaults file.
                 rep_user -> Replication user name.
-                rep_japd -> Replication user password.
+                rep_japd -> Replication user pswd.
                 host -> Host name or IP of server.
                 port -> Port for MySQL.
                 defaults_file -> Location of my.cnf file.
@@ -1523,7 +1586,7 @@ class SlaveRep(Rep):
             (input) name -> Name of the MySQL server.
             (input) server_id -> Server's ID.
             (input) sql_user -> SQL user's name.
-            (input) sql_pass -> SQL user's password.
+            (input) sql_pass -> SQL user's pswd.
             (input) os_type -> Machine operating system type class instance.
             (input) host -> 'localhost' or host name or IP.
             (input) port -> '3306' or port for MySQL.
@@ -1534,7 +1597,7 @@ class SlaveRep(Rep):
                 port -> Port for MySQL.
                 defaults_file -> Location of my.cnf file.
                 rep_user -> Replication user name.
-                rep_japd -> Replication user password.
+                rep_japd -> Replication user pswd.
                 ssl_client_ca -> SSL certificate authority file.
                 ssl_client_key -> SSL X.509 key file.
                 ssl_client_cert -> SSL X.509 certificate file.
@@ -1656,16 +1719,20 @@ class SlaveRep(Rep):
 
         """
 
+        # Semantic change in MySQL 8.0.22
+        master = "Source" if self.version >= (8, 0, 22) else "Master"
+        slave = "Replica" if self.version >= (8, 0, 22) else "Slave"
+
         slave_stop(self)
         data = show_slave_stat(self)[0]
 
-        self.io_state = data["Slave_IO_State"]
+        self.io_state = data[slave + "_IO_State"]
 
         try:
-            self.secs_behind = int(data["Seconds_Behind_Master"])
+            self.secs_behind = int(data["Seconds_Behind_" + master])
 
         except (ValueError, TypeError):
-            self.secs_behind = data["Seconds_Behind_Master"]
+            self.secs_behind = data["Seconds_Behind_" + master]
 
     def start_slave(self):
 
@@ -1678,16 +1745,20 @@ class SlaveRep(Rep):
 
         """
 
+        # Semantic change in MySQL 8.0.22
+        master = "Source" if self.version >= (8, 0, 22) else "Master"
+        slave = "Replica" if self.version >= (8, 0, 22) else "Slave"
+
         slave_start(self)
         data = show_slave_stat(self)[0]
 
-        self.io_state = data["Slave_IO_State"]
+        self.io_state = data[slave + "_IO_State"]
 
         try:
-            self.secs_behind = int(data["Seconds_Behind_Master"])
+            self.secs_behind = int(data["Seconds_Behind_" + master])
 
         except (ValueError, TypeError):
-            self.secs_behind = data["Seconds_Behind_Master"]
+            self.secs_behind = data["Seconds_Behind_" + master]
 
     def show_slv_state(self):
 
@@ -1714,10 +1785,13 @@ class SlaveRep(Rep):
 
         """
 
+        # Semantic change in MySQL 8.0.22
+        slave = "Replica" if self.version >= (8, 0, 22) else "Slave"
+
         data = show_slave_stat(self)[0]
-        self.io_state = data["Slave_IO_State"]
-        self.slv_io = data["Slave_IO_Running"]
-        self.slv_sql = data["Slave_SQL_Running"]
+        self.io_state = data[slave + "_IO_State"]
+        self.slv_io = data[slave + "_IO_Running"]
+        self.slv_sql = data[slave + "_SQL_Running"]
 
     def upd_slv_status(self):
 
@@ -1729,18 +1803,26 @@ class SlaveRep(Rep):
 
         """
 
+        # Semantic change in MySQL 8.0.22
+        master = "Source" if self.version >= (8, 0, 22) else "Master"
+        slave = "Replica" if self.version >= (8, 0, 22) else "Slave"
+
+        # Semantic change in MySQL 8.0.26
+        slave2 = "replica" if self.version >= (8, 0, 26) else "slave"
+        slave3 = "Replica" if self.version >= (8, 0, 26) else "Slave"
+
         data = show_slave_stat(self)[0]
-        self.io_state = data["Slave_IO_State"]
-        self.mst_host = data["Master_Host"]
-        self.mst_port = data["Master_Port"]
+        self.io_state = data[slave + "_IO_State"]
+        self.mst_host = data[master + "_Host"]
+        self.mst_port = data[master + "_Port"]
         self.conn_retry = data["Connect_Retry"]
-        self.mst_log = data["Master_Log_File"]
-        self.mst_read_pos = data["Read_Master_Log_Pos"]
+        self.mst_log = data[master + "_Log_File"]
+        self.mst_read_pos = data["Read_" + master + "_Log_Pos"]
         self.relay_log = data["Relay_Log_File"]
         self.relay_pos = data["Relay_Log_Pos"]
-        self.relay_mst_log = data["Relay_Master_Log_File"]
-        self.slv_io = data["Slave_IO_Running"]
-        self.slv_sql = data["Slave_SQL_Running"]
+        self.relay_mst_log = data["Relay_" + master + "_Log_File"]
+        self.slv_io = data[slave + "_IO_Running"]
+        self.slv_sql = data[slave + "_SQL_Running"]
         self.do_db = data["Replicate_Do_DB"]
         self.ign_db = data["Replicate_Ignore_DB"]
         self.do_tbl = data["Replicate_Do_Table"]
@@ -1756,25 +1838,25 @@ class SlaveRep(Rep):
         except ValueError:
             self.skip_ctr = data["Skip_Counter"]
 
-        self.exec_mst_pos = data["Exec_Master_Log_Pos"]
+        self.exec_mst_pos = data["Exec_" + master + "_Log_Pos"]
         self.log_space = data["Relay_Log_Space"]
         self.until_cond = data["Until_Condition"]
         self.until_log = data["Until_Log_File"]
         self.until_pos = data["Until_Log_Pos"]
-        self.ssl_allow = data["Master_SSL_Allowed"]
-        self.ssl_file = data["Master_SSL_CA_File"]
-        self.ssl_path = data["Master_SSL_CA_Path"]
-        self.ssl_cert = data["Master_SSL_Cert"]
-        self.ssl_cipher = data["Master_SSL_Cipher"]
-        self.ssl_key = data["Master_SSL_Key"]
+        self.ssl_allow = data[master + "_SSL_Allowed"]
+        self.ssl_file = data[master + "_SSL_CA_File"]
+        self.ssl_path = data[master + "_SSL_CA_Path"]
+        self.ssl_cert = data[master + "_SSL_Cert"]
+        self.ssl_cipher = data[master + "_SSL_Cipher"]
+        self.ssl_key = data[master + "_SSL_Key"]
 
         try:
-            self.secs_behind = int(data["Seconds_Behind_Master"])
+            self.secs_behind = int(data["Seconds_Behind_" + master])
 
         except (ValueError, TypeError):
-            self.secs_behind = data["Seconds_Behind_Master"]
+            self.secs_behind = data["Seconds_Behind_" + master]
 
-        self.ssl_verify = data["Master_SSL_Verify_Server_Cert"]
+        self.ssl_verify = data[master + "_SSL_Verify_Server_Cert"]
 
         try:
             self.io_err = int(data["Last_IO_Errno"])
@@ -1794,28 +1876,28 @@ class SlaveRep(Rep):
         self.ign_ids = data["Replicate_Ignore_Server_Ids"]
 
         try:
-            self.mst_id = int(data["Master_Server_Id"])
+            self.mst_id = int(data[master + "_Server_Id"])
 
         except ValueError:
-            self.mst_id = data["Master_Server_Id"]
+            self.mst_id = data[master + "_Server_Id"]
 
-        self.mst_uuid = data.get("Master_UUID", None)
-        self.mst_info = data.get("Master_Info_File", None)
+        self.mst_uuid = data.get(master + "_UUID", None)
+        self.mst_info = data.get(master + "_Info_File", None)
         self.sql_delay = data.get("SQL_Delay", None)
         self.sql_remain = data.get("SQL_Remaining_Delay", None)
-        self.slv_sql_state = data.get("Slave_SQL_Running_State", None)
-        self.mst_retry = data.get("Master_Retry_Count", None)
-        self.mst_bind = data.get("Master_Bind", None)
+        self.slv_sql_state = data.get(slave + "_SQL_Running_State", None)
+        self.mst_retry = data.get(master + "_Retry_Count", None)
+        self.mst_bind = data.get(master + "_Bind", None)
         self.io_err_time = data.get("Last_IO_Error_Timestamp", None)
         self.sql_err_time = data.get("Last_SQL_Error_Timestamp", None)
-        self.ssl_crl = data.get("Master_SSL_Crl", None)
-        self.ssl_crl_path = data.get("Master_SSL_Crlpath", None)
+        self.ssl_crl = data.get(master + "_SSL_Crl", None)
+        self.ssl_crl_path = data.get(master + "_SSL_Crlpath", None)
         self.retrieved_gtid = data.get("Retrieved_Gtid_Set", None)
         self.exe_gtid = data.get("Executed_Gtid_Set", None)
         self.auto_pos = data.get("Auto_Position", None)
 
-        # tran_retry and run are in different location in MySQL 8.0
-        if self.version[0] < 8:
+        # tran_retry and run are in different location in MySQL 8
+        if self.version < (8, 0, 0):
             self.run = fetch_global_var(self, "slave_running")["Slave_running"]
             self.tran_retry = fetch_global_var(
                 self,
@@ -1830,7 +1912,7 @@ class SlaveRep(Rep):
             self.tran_retry = self.col_sql(sql % (ctr))[0][ctr]
 
         self.tmp_tbl = fetch_global_var(
-            self, "slave_open_temp_tables")["Slave_open_temp_tables"]
+            self, slave2 + "_open_temp_tables")[slave3 + "_open_temp_tables"]
         self.read_only = fetch_sys_var(self, "read_only")["read_only"]
 
         self.upd_gtid_pos()
@@ -1964,13 +2046,16 @@ class SlaveRep(Rep):
 
         """
 
+        # Semantic change in MySQL 8.0.22
+        master = "Source" if self.version >= (8, 0, 22) else "Master"
+
         data = show_slave_stat(self)[0]
 
         try:
-            self.secs_behind = int(data["Seconds_Behind_Master"])
+            self.secs_behind = int(data["Seconds_Behind_" + master])
 
         except (ValueError, TypeError):
-            self.secs_behind = data["Seconds_Behind_Master"]
+            self.secs_behind = data["Seconds_Behind_" + master]
 
     def get_time(self):