+++ test/integration/mysql_libs/find_name.py
@@ -29,7 +29,7 @@ else:
 sys.path.append(os.getcwd())
 import mysql_libs
 import mysql_class
-import lib.cmds_gen as cmds_gen
+import lib.gen_libs as gen_libs
 import version
 
 __version__ = version.__version__
@@ -73,7 +73,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        slaves = cmds_gen.create_cfg_array(self.config_name,
+        slaves = gen_libs.create_cfg_array(self.config_name,
                                            cfg_path=self.config_dir)
         servers = mysql_libs.create_slv_array(slaves)
         slv = mysql_libs.find_name(servers, self.name)
@@ -90,7 +90,7 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        slaves = cmds_gen.create_cfg_array(self.config_name,
+        slaves = gen_libs.create_cfg_array(self.config_name,
                                            cfg_path=self.config_dir)
         name = slaves[0]["name"]
         servers = mysql_libs.create_slv_array(slaves)