+++ test/unit/mysql_class/slaverep_updslvstate.py
@@ -43,6 +43,8 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_post_8022
+        test_pre_8022
         test_value
 
     """
@@ -67,9 +69,58 @@ class UnitTest(unittest.TestCase):
         self.defaults_file = "def_cfg_file"
         self.extra_def_file = "extra_cfg_file"
 
-        self.show_stat = [{"Slave_IO_State": "Up",
-                           "Slave_IO_Running": "Running",
-                           "Slave_SQL_Running": "SQL_Code"}]
+        self.show_stat = [
+            {"Slave_IO_State": "Up", "Slave_IO_Running": "Running",
+             "Slave_SQL_Running": "SQL_Code"}]
+        self.show_stat2 = [
+            {"Replica_IO_State": "Up", "Replica_IO_Running": "Running",
+             "Replica_SQL_Running": "SQL_Code"}]
+
+    @mock.patch("mysql_class.show_slave_stat")
+    def test_post_8022(self, mock_stat):
+
+        """Function:  test_post_8022
+
+        Description:  Test with post-MySQL 8.0.22.
+
+        Arguments:
+
+        """
+
+        mock_stat.return_value = self.show_stat2
+
+        mysqlrep = mysql_class.SlaveRep(
+            self.name, self.server_id, self.sql_user, self.sql_pass,
+            self.machine, defaults_file=self.defaults_file)
+        mysqlrep.version = (8, 0, 28)
+        mysqlrep.upd_slv_state()
+
+        self.assertEqual(
+            (mysqlrep.io_state, mysqlrep.slv_io, mysqlrep.slv_sql),
+            ("Up", "Running", "SQL_Code"))
+
+    @mock.patch("mysql_class.show_slave_stat")
+    def test_pre_8022(self, mock_stat):
+
+        """Function:  test_pre_8022
+
+        Description:  Test with pre-MySQL 8.0.22.
+
+        Arguments:
+
+        """
+
+        mock_stat.return_value = self.show_stat
+
+        mysqlrep = mysql_class.SlaveRep(
+            self.name, self.server_id, self.sql_user, self.sql_pass,
+            self.machine, defaults_file=self.defaults_file)
+        mysqlrep.version = (8, 0, 21)
+        mysqlrep.upd_slv_state()
+
+        self.assertEqual(
+            (mysqlrep.io_state, mysqlrep.slv_io, mysqlrep.slv_sql),
+            ("Up", "Running", "SQL_Code"))
 
     @mock.patch("mysql_class.show_slave_stat")
     def test_value(self, mock_stat):
@@ -83,15 +134,16 @@ class UnitTest(unittest.TestCase):
         """
 
         mock_stat.return_value = self.show_stat
-        mysqlrep = mysql_class.SlaveRep(self.name, self.server_id,
-                                        self.sql_user, self.sql_pass,
-                                        self.machine,
-                                        defaults_file=self.defaults_file)
 
+        mysqlrep = mysql_class.SlaveRep(
+            self.name, self.server_id, self.sql_user, self.sql_pass,
+            self.machine, defaults_file=self.defaults_file)
+        mysqlrep.version = (8, 0, 21)
         mysqlrep.upd_slv_state()
-        self.assertEqual((mysqlrep.io_state, mysqlrep.slv_io,
-                          mysqlrep.slv_sql),
-                         ("Up", "Running", "SQL_Code"))
+
+        self.assertEqual(
+            (mysqlrep.io_state, mysqlrep.slv_io, mysqlrep.slv_sql),
+            ("Up", "Running", "SQL_Code"))
 
 
 if __name__ == "__main__":