+++ test/unit/mysql_libs/code_coverage.sh
@@ -27,7 +27,6 @@ coverage run -a --source=mysql_libs test/unit/mysql_libs/reset_slave.py
 coverage run -a --source=mysql_libs test/unit/mysql_libs/select_wait_until.py
 coverage run -a --source=mysql_libs test/unit/mysql_libs/create_slv_array.py
 coverage run -a --source=mysql_libs test/unit/mysql_libs/crt_cmd.py
-coverage run -a --source=mysql_libs test/unit/mysql_libs/crt_srv_inst.py
 coverage run -a --source=mysql_libs test/unit/mysql_libs/fetch_slv.py
 coverage run -a --source=mysql_libs test/unit/mysql_libs/find_name.py
 coverage run -a --source=mysql_libs test/unit/mysql_libs/is_cfg_valid.py