+++ test/unit/mysql_libs/unit_test_run.sh
@@ -24,7 +24,6 @@ test/unit/mysql_libs/reset_slave.py
 test/unit/mysql_libs/select_wait_until.py
 test/unit/mysql_libs/create_slv_array.py
 test/unit/mysql_libs/crt_cmd.py
-test/unit/mysql_libs/crt_srv_inst.py
 test/unit/mysql_libs/fetch_slv.py
 test/unit/mysql_libs/find_name.py
 test/unit/mysql_libs/is_cfg_valid.py